// frontend/my-app/src/models/FlightFilterOptions.ts
export enum FlightFilterOptions {
    GET_ALL_FLIGHTS = 'get_all_flights',
    GET_FLIGHT_BY_ID = 'get_flight_by_id',
    GET_FLIGHTS_BY_AIRLINE_ID = 'get_flights_by_airline_id',
    GET_FLIGHTS_BY_PARAMETERS = 'get_flights_by_parameters',
    GET_ARRIVAL_FLIGHTS = 'get_arrival_flights',
    GET_DEPARTURE_FLIGHTS = 'get_departure_flights',
  }
  